feat(worker/mail): rename `verifyEmail` function to `sendAccountVerificationEmail`
- Use in auth/email/verify and auth/createUser
- Move logic to mail/worker.service.ts file
